Completely changed data structure
Wasn't a big fan of the way it was done previously. If you need an
explanation, the new system is this: the plugin stores to memory a data
source, of the interface DataSource. When needed, the plugin will call
the data source's methods. There are two implementations of DataSource:
FlatFIleDataSource, and MySQLDataSource. The first implementation saves
the player list to players.yml, whereas the latter implementation saves
the player list to the database. When the plugin is loaded, it checks
config.yml to see if it should use MySQL, and if so, it'll go ahead and
set its data source to a new MySQLDataSource, and in doing that, it
sets up the connection and everything. If that fails, or if
mysql.enabled is false, the plugin will set its data source to a new
FlatFileDataSource, which will create a new file called players.yml and
load that up.

It shouldn't be too complicated, and IMO it looks better, and
structurally, it's better than what we had before.